[CHORE] collection view 변수명 table view로 수정(#80)
